requirements.txt
streamlit
google-generativeai
Pillow
reportlab
python-dotenv